character now rotates correctly and the projectiles follow his rotation

+++ SluaghEngine/includes/Gameplay/PlayerUnit.h
@@ -66,8 +66,8 @@ namespace SE
 				bool downD;
 				
 				bool mouseRightDown;
-				int mousePosX;
-				int mousePosY;
+				float mousePosX;
+				float mousePosY;
 
 				MovementInput()
 				{
